FOUNDATION-11: implement data layer

+++ feature/home/src/main/java/ru/appricot/startuphub/home/details/StartupDetailsScreen.kt
@@ -24,7 +24,7 @@ fun StartupDetailsScreen(viewModel: StartupDetailsViewModel, onBackClick: () -> 
 }
 
 @Composable
-fun Content(id: Long?, onButtonClick: () -> Unit, modifier: Modifier = Modifier) {
+fun Content(id: Int?, onButtonClick: () -> Unit, modifier: Modifier = Modifier) {
     Column(
         modifier = modifier
             .fillMaxSize()
